+++ Source/cmMakefile.cxx
@@ -322,8 +322,8 @@ bool cmMakefile::ReadListFile(const char* filename, const char* external)
     }
   // add this list file to the list of dependencies
   m_ListFiles.push_back( filenametoread);
-  const int numberFunctions = lf->m_Functions.size();
-  for(int i =0; i < numberFunctions; ++i)
+  const size_t numberFunctions = lf->m_Functions.size();
+  for(size_t i =0; i < numberFunctions; ++i)
     {
     cmListFileFunction& curFunction = lf->m_Functions[i];
     if(!this->IsFunctionBlocked(curFunction.m_Name.c_str(),
